revert: reverted bad merge on develop Refs: d54bed0, c482581

+++ PERUSTARS/PERUSTARS/Shared/Domain/Repositories/IBaseRepository.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace PERUSTARS.Shared.Domain.Repositories
@@ -7,7 +6,7 @@ namespace PERUSTARS.Shared.Domain.Repositories
     public interface IBaseRepository<TEntity>
     {
         Task AddAsync(TEntity entity);
-        Task<TEntity?> FindByIdAsync(long id);
+        Task<TEntity?> FindByIdAsync(int id);
         void Update(TEntity entity);
         void Remove(TEntity entity);
         Task<IEnumerable<TEntity>> ListAsync();
