+++ PERUSTARS/PERUSTARS/Shared/Infrastructure/Repositories/BaseRepository.cs
@@ -2,7 +2,6 @@
 using PERUSTARS.Shared.Domain.Repositories;
 using PERUSTARS.Shared.Infrastructure.Configuration;
 using System.Collections.Generic;
-using System.Numerics;
 using System.Threading.Tasks;
 
 namespace PERUSTARS.Shared.Infrastructure.Repositories
@@ -19,7 +18,7 @@ namespace PERUSTARS.Shared.Infrastructure.Repositories
             await _dbContext.Set<TEntity>().AddAsync(entity);
         }
 
-        public async Task<TEntity?> FindByIdAsync(long id)
+        public async Task<TEntity?> FindByIdAsync(int id)
         {
             return await _dbContext.Set<TEntity>().FindAsync(id);
         }
